+++ Sources/StreamChat/ChatClient.swift
@@ -90,6 +90,12 @@ public class ChatClient {
 
     let channelListUpdater: ChannelListUpdater
 
+    /// Handler for watching channels and preventing duplicate watch requests.
+    ///
+    /// If a channel is created and belongs to multiple queries at the same time,
+    /// we want to make sure we only watch it one time, not for every query it belongs.
+    let channelWatcherHandler: ChannelWatcherHandling
+
     func makeMessagesPaginationStateHandler() -> MessagesPaginationStateHandling {
         MessagesPaginationStateHandler()
     }
@@ -218,6 +224,9 @@ public class ChatClient {
             apiClient
         )
         pollsRepository = environment.pollsRepositoryBuilder(databaseContainer, apiClient)
+        channelWatcherHandler = ChannelWatcherHandler(
+            channelListUpdater: channelListUpdater
+        )
 
         authRepository.delegate = self
         apiClientEncoder.connectionDetailsProviderDelegate = self